Fix violation of ESLint rule `no-constant-binary-expression`
This one is a bit tricky, because it also removes a call of `this.R` which is an alias for `getRandomFloat` from `./renderengine/helper.js` — which technically has a side-effect, because it generates a new deterministic random number. But this should be fine there, because `0.8` is truthy and therefor `this.R` will never be called in this line.