Fix violation of ESLint rule `no-sparse-arrays`
This one is tricky. Technically a sparse array like [,1] will result in an array of `length` `2` that looks like `[empty,1]`, which might result in a different runtime behavior then just `[1]`. But we know these arrays get filtered and falsy elements will be ignored, so this change is save.

In the previous commit a snapshot test was also added, to ensure this doesn't result in an unexpected regression.