Enable ESLint rule `no-unused-vars`
This enables the recommended ESLint rule `no-unused-vars`, but also
overwrites it right away with the `@typescript-eslint` version of the
same rule, which respects some TypeScript specifics of wether or not a
variable is unused.

This also adds some exception, where an unused variable is followed
by a used parameter, in which case fixing this rule would mean
changing the function signature.