Fix violation of `no-unused-vars`
This one is a bit tricky, because the `nextFirstColor` and
`nextSecondColor` are calling `this.IF` which is an alias for `getIf`
from `./renderengine/helpers.js`, which has the side-effect of
generating a new random number. As a result all the following random
numbers will also be different — which in this case drastically changes
the generated image.

This change is acceptable though, because changing the generated image
relative to the seed is not considered a breaking change (by me).

It does however mean that the snapshot test will have to be updated.

See the PR message for #22 for more details.